requirements.txt
Django==5.2
djangorestframework==3.16.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
